matriz: Literature Matrix Synthesis Tools for Epidemiology and Health Science Research

An easy-to-use workflow that provides tools to create, update and fill literature matrices commonly used in research, specifically epidemiology and health sciences research. The project is born out of need as an easy–to–use tool for my research methods classes.

See also

Author

Maintainer: JP Monteagudo jpmonteagudo2014@gmail.com (ORCID) [copyright holder]