Display package version for matriz

matriz_message() produces a message about the package version and the version of R making use of this package.

Usage

matriz_message()

Value

dmatriz_message() returns a message about the install version of matriz.

Author

JP Monteagudo

Examples

matriz_message()
#> -------------------------
#> matriz string generated by {matriz} v.1.0.1.
#> R version 4.4.2 (2024-10-31).
#> -------------------------
#> Easily generate and manage structured literature review matrices in R.
#> Streamline your research synthesis, track key study details, and organize citations efficiently.